requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1


certifi==2021.5.30
charset-normalizer==2.0.4

idna==3.2


requests==2.26.0
six==1.16.0


urllib3==1.26.6
gunicorn==20.1.0
whitenoise==5.3.0
